Add files via upload
I wrote a single page application that has a form with a  textbox for your full name, radio buttons for gender, and a drop‐down for favorite day of  the week. I used HTML5 validation for these fields.  There are two buttons, one labelled submit and another one labelled clear.  The clear button resets button and clear the form.  The other button is of type submit and when pressed it calls a JavaScript function that will return false.   Inside of my validation function I alerted the user what their full name, gender, and favorite day is.